feat: Implement AddEntityToCollectionsForm component for managing entity collections
